+++ pochidetection/utils/__init__.py
@@ -1,5 +1,10 @@
 """ユーティリティモジュール."""
 
+from pochidetection.utils.benchmark import (
+    BenchmarkResult,
+    build_benchmark_result,
+    write_benchmark_result,
+)
 from pochidetection.utils.config_loader import ConfigLoader
 from pochidetection.utils.history import TrainingHistory
 from pochidetection.utils.phased_timer import PhasedTimer
@@ -7,9 +12,12 @@ from pochidetection.utils.timer import InferenceTimer
 from pochidetection.utils.work_dir import WorkspaceManager
 
 __all__ = [
+    "BenchmarkResult",
     "ConfigLoader",
     "InferenceTimer",
     "PhasedTimer",
     "TrainingHistory",
     "WorkspaceManager",
+    "build_benchmark_result",
+    "write_benchmark_result",
 ]